DW-3020 move files to matching packages, remname non-legacy writer
